Update All.ino
Updates with Raiym. (Last commit I accidentally pasted the new code without deleting the old)
Everything works except for case 1 scroll text - that code works on its own, the rest of the code works on its own, but together they make case 1 glitch/freeze the device. We isolated the bug to a single line (matrix.show) and commented that out for now.
Once that bug is fixed it can be implemented for case 2 and we are done.